Add of new USART-all function to deal with "string"
