--- 01_Code_Source-EMB/FrogINT_source/Debug/FrogINT.hex
@@ -1,56 +1,0 @@
-:100000000C942A000C9447000C9447000C94470071
-:100010000C9447000C9447000C9447000C94470044
-:100020000C9447000C9447000C9447000C94470034
-:100030000C9447000C9447000C9447000C94470024
-:100040000C9447000C9447000C9447000C94470014
-:100050000C94470011241FBECFE5D4E0DEBFCDBF16
-:1000600011E0A0E6B0E0E0ECF2E002C005900D92F5
-:10007000A031B107D9F721E0A0E1B1E001C01D92A4
-:10008000AC34B207E1F70E94FF000C945E010C94BF
-:1000900000008FEF87BB0895009721F0019721F4AE
-:1000A000C09A0895C0980895C0980895AC01CB01F6
-:1000B000880F991F680F791F3BB33E7F2BB32160D8
-:1000C0009AB3922B9ABBAFB7F8946115710589F07A
-:1000D000FA01640F751F919188E02BBB97FF3BBB22
-:1000E000990F000000C03BBB8A95B9F7E617F707E8
-:1000F00091F7AFBF85E88A95F1F70000089590BDAC
-:1001000089B988E18AB98EE880BD08955D9BFECFEC
-:100110008CB90895CF93DF93EC018881882331F067
-:1001200021960E94860089918111FBCFDF91CF91AA
-:1001300008955F9BFECF8CB10895CF930E949900E4
-:10014000C82F90E00E945201892B11F4C03239F47B
-:100150008C2F0E94860080E690E00E9486008C2F03
-:10016000CF910895CF93C82F8823A1F06C2F70E012
-:10017000FB013197CF01880F991FE80FF91FE85D48
-:10018000FE4F11821082128288E291E00E94560096
-:10019000C15061F72FEF84E39CE0215080409040F4
-:1001A000E1F700C00000CF910895E9E4F1E025E215
-:1001B00031E08FEF8183108212823397E217F307C9
-:1001C000C9F70895E9E4F1E025E231E08FEF11820B
-:1001D000808312823397E217F307C9F7089564E228
-:1001E00070E088E291E00E9456002FEF89E698E1E6
-:1001F000215080409040E1F700C0000008958CE05D
-:100200000E94B2000E94D5000E94E20083E390E0C9
-:100210000E947F000E94490083EF90E00E948A00C4
-:1002200086ED90E00E948A0084EC90E00E948A00B3
-:1002300087EA90E00E948A0083EF90E00E948A00A3
-:100240000E94EF008EE990E00E948A000E949D00CB
-:10025000C82F883619F08F3671F019C081E090E010
-:100260000E944C000E94D5000E94EF0081E890E0BF
-:100270000E948A00E5CF80E090E00E944C000E943E
-:10028000E2000E94EF0083E690E00E948A008C2F3B
-:1002900090E00E945201892B99F283EA90E00E943B
-:1002A0008A00CECF803209F091110C945B0180520C
-:1002B0008F55D8F70895992788270895F894FFCF88
-:1002C0000D0A004C454420506F727442302069730F
-:1002D000207377697463686564204F4646210D0A70
-:1002E000004C454420506F72744230206973207373
-:1002F00077697463686564204F4E210D0A000D0A0A
-:100300003E2000082008002A2A2A2A2A2A2A56657E
-:100310007273696F6E2056302E312A2A2A2A2A2AB1
-:100320002A0D0A002A2A2A2A46726F67494E542A41
-:100330002A2A2A0D0A002A2A2A2A466C6F72656E1A
-:1003400074205461696E7475726965722A2A2A2A4A
-:100350000D0A002A2A2A2A2A2A2A2A2A2A2A2A2A64
-:100360002A2A2A2A2A2A2A2A2A2A2A2A2A0D0A0054
-:00000001FF
